requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
neo4j==5.14.0
pandas==2.1.3
arxiv==2.0.0
ollama==0.1.6